mobile/requirements.txt
Flask-Uploads==0.1.3
